Rename NewPageHeader* components to PageHeader*
